refactor: Sections 일급 컬렉션으로 분리
